fix [parser]: read table to gen data

+++ src/core/schema.rs
@@ -25,9 +25,9 @@ pub struct ColumnSchema {
     pub is_nullable: bool,
     pub is_primary_key: bool,
     pub is_auto_increment: bool,
-    pub max_length: Option<i32>,
-    pub numeric_precision: Option<i32>,
-    pub numeric_scale: Option<i32>,
+    pub max_length: Option<i64>,
+    pub numeric_precision: Option<i64>,
+    pub numeric_scale: Option<i64>,
     pub default_value: Option<String>,
     pub is_unique: bool,
 }
